workers/requirements.txt
# Celery & Task Queue
celery==5.3.6
redis==5.0.1
kombu==5.3.5

# Database
sqlalchemy==2.0.25
psycopg2-binary==2.9.9
asyncpg==0.29.0
motor==3.3.2
pymongo==4.6.1

# HTTP Client
httpx==0.26.0
aiohttp==3.9.1

# Web Scraping
playwright==1.41.0
beautifulsoup4==4.12.3
lxml==5.1.0

# Data Processing
pydantic==2.5.3
pydantic-settings==2.1.0
orjson==3.9.12

# Utilities
python-dotenv==1.0.0
tenacity==8.2.3

# Monitoring
prometheus-client==0.19.0
sentry-sdk==1.39.1
flower==2.0.1  # Celery monitoring

# Social Media (unofficial APIs - use carefully)
# TikTokApi==6.1.1  # May need manual setup

